==> code/1_scraping.ipynb <==
    The purpose of this notebook is to collect Reddit.com posts from two 'subreddits': AskWomen and AskMen.

import requests
import json
import time
import pandas as pd

w_posts = []
after = None
params={}
for i in range(150):
    url='https://www.reddit.com/r/AskWomen.json'
    w_requests = requests.get(url, params=params, headers={'user-agent': 'seany boy'})
    if w_requests.status_code == 200:
        the_json = w_requests.json()
        w_posts.extend(the_json['data']['children'])
        after = the_json['data']['after']
        params = {'after': the_json['data']['after']}
    else:
        print(w_requests.status_code)
    if (i+1)%15 == 0:
        print(f'{(i+1)/1.5}%')
    else:
        pass
    time.sleep(1)
    
#This code was inspired by Riley Dallas' video titled "Project 3 Info"

m_posts = []
after = None
params={}
for i in range(150):
    url='https://www.reddit.com/r/AskMen.json'
    m_requests = requests.get(url, params=params, headers={'user-agent': 'seany boy'})
    if m_requests.status_code == 200:
        the_json = m_requests.json()
        m_posts.extend(the_json['data']['children'])
        after = the_json['data']['after']
        params = {'after': the_json['data']['after']}
    else:
        print(m_requests.status_code)
    if (i+1)%15 == 0:
        print(f'{(i+1)/1.5}%')
    else:
        pass
    time.sleep(1)
    
#This code was inspired by Riley Dallas' video titled "Project 3 Info"

w_ids = list(set([w_posts[i]['data']['id'] for i in range(len(w_posts))]))
w_list = []
for i in range(len(w_posts)):
    id_ = w_posts[i]['data']['id']
    if id_ in w_ids:
        w_dict = {}
        w_dict['title'] = w_posts[i]['data']['title']
        w_dict['text'] = w_posts[i]['data']['selftext']
        w_dict['source'] = 'AskWomen'
        w_list.append(w_dict)
        w_ids.remove(id_)
    else:
        pass

m_ids = list(set([m_posts[i]['data']['id'] for i in range(len(m_posts))]))
m_list = []
for i in range(len(m_posts)):
    id_ = m_posts[i]['data']['id']
    if id_ in m_ids:
        m_dict = {}
        m_dict['title'] = m_posts[i]['data']['title']
        m_dict['text'] = m_posts[i]['data']['selftext']
        m_dict['source'] = 'AskMen'
        m_list.append(m_dict)
        m_ids.remove(id_)
    else:
        pass

w_df = pd.DataFrame(w_list)
m_df = pd.DataFrame(m_list)
print(w_df.shape)
print(m_df.shape)

both_df = pd.DataFrame(w_list + m_list)
print(both_df.shape)

# both_df.to_csv('../data/data.csv')

    Bottom line commented out to avoid overwriting "data".
