==> code/2_cleaning.ipynb <==
    In this notebook, we remove unwanted characters, make all letters lowercase, apply PorterStemmer to each
    post title. We create a target column where 1 indicates the post is from AskWomen and 0 from AskMen.

import pandas as pd
import regex as re
from nltk.stem.porter import PorterStemmer
from nltk.corpus import stopwords

df = pd.read_csv('../data/data.csv')
df.columns

df = df.drop(columns=['Unnamed: 0', 'index'])

df.isna().sum()

df['source'].value_counts(normalize=True)

df.shape

df['char_count'] = df['title'].apply(len)
df['word_count'] = df['title'].apply(lambda x: len(x.split()))
# Code adopted from Roy Kim's lesson on NLP

df.groupby(by=['source']).mean()

df['title'].tail()

df['title'][1265]

df['title'] = [str(x).replace('&amp;', '&').replace('\n', '').replace("\'", "'") for x in df['title']]

df['title'] = [str(x).replace("SO", "sigoth").replace("significant other", "sigoth") for x in df['title']]

p_stemmer = PorterStemmer()
clean_titles = []
for title in df['title']:
    lower_letters = re.sub("[^a-zA-Z]", " ", title).replace("  ", " ").replace('   ', ' ')
    no_stops = [w.lower() for w in lower_letters.split() if w not in stopwords.words('english')]
    stemmed = [p_stemmer.stem(word) for word in no_stops]
    clean_titles.append(" ".join(stemmed))

len(clean_titles)

df['clean_titles'] = pd.Series(clean_titles)

df['clean_titles'].tail()

df = pd.get_dummies(df, columns=['source'], drop_first=True)

df['target'] = df['source_AskWomen']

df = df.drop(columns='source_AskWomen')

df.to_csv('../data/clean_data.csv', index=False)
