==> code/3_NLP_EDA.ipynb <==
    In this notebook, we create data frames with both CountVectorizer and TfidfVectorizer. We also explore the
    words which appear in one subreddit and not the other, as well as words most correlated with the target.

import pandas as pd
import numpy as np
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer

df = pd.read_csv('../data/clean_data.csv').drop(columns=['text', 'title'])

df.head()

cvec = CountVectorizer(stop_words='english', min_df=3, max_df=1.0, ngram_range=(1,2))
cvec_df = pd.SparseDataFrame(cvec.fit_transform(df['clean_titles']).toarray(),
                           columns=cvec.get_feature_names()).join(df['target'])

cvec_df.head()

tfidf = TfidfVectorizer(stop_words='english', min_df=3, max_df=1.0, ngram_range=(1,2))
tfidf_df = pd.SparseDataFrame(tfidf.fit_transform(df['clean_titles']).toarray(),
                           columns=tfidf.get_feature_names()).join(df['target'])

cvec_df.to_csv('../data/cvec_df.csv', index=False)
tfidf_df.to_csv('../data/tfidf_df.csv', index=False)

term_dict = {word : cvec_df[word].sum() for word in cvec_df.drop('target', axis=1).columns}

tcdf = pd.DataFrame(term_dict, index = ['counts']).T

top_women = list(cvec_df.groupby('target').sum().T.sort_values(1, ascending=False).head(50).index)
top_men = list(cvec_df.groupby('target').sum().T.sort_values(0, ascending=False).head(50).index)

women_only = [word for word in top_women if word not in top_men]
women_only

men_only = [word for word in top_men if word not in top_women]
men_only

overlap = [word for word in top_men if word in top_women]
overlap

cvec_df.corr()[['target']].sort_values('target', ascending=False).head(10)

cvec_df.corr()[['target']].sort_values('target', ascending=True).head(10)

top_women = list(tfidf_df.groupby('target').sum().T.sort_values(1, ascending=False).head(50).index)
top_men = list(tfidf_df.groupby('target').sum().T.sort_values(0, ascending=False).head(50).index)

women_only = [word for word in top_women if word not in top_men]
women_only

men_only = [word for word in top_men if word not in top_women]
men_only

overlap = [word for word in top_men if word in top_women]
overlap

tfidf_df.corr()[['target']].sort_values('target', ascending=False).head(10)

tfidf_df.corr()[['target']].sort_values('target', ascending=True).head(10)
