==> code/5_modeling_tfidf.ipynb <==
    In this notebook, we model our TFIDF data with Logistic Regression and Multinomial Naive Bayes.

import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import GridSearchCV
from sklearn.naive_bayes import MultinomialNB
from sklearn.metrics import accuracy_score

df = pd.read_csv('../data/tfidf_df.csv')
clean_df = pd.read_csv('../data/clean_data.csv')

df.head()

X = df.drop(columns='target')
y = clean_df['target']

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=7, stratify=y)

lr = LogisticRegression(random_state=7)

params = {'penalty': ['l1'],
          'solver': ['liblinear'],
          'C': list(np.logspace(-0.1, 0.1, 20)),
         'max_iter': [100, 150]}

gs = GridSearchCV(lr, param_grid=params, cv=5)

gs.fit(X_train, y_train)

gs.best_params_

gs.best_score_

y_pred_train = gs.predict(X_train)
y_pred_test = gs.predict(X_test)

train_score = accuracy_score(y_train, y_pred_train)
test_score = accuracy_score(y_test, y_pred_test)

print(f"Train accuracy: {round(train_score, 4)}")
print(f"Test accuracy: \t{round(test_score, 4)}")

mnb = MultinomialNB()

params = {'alpha': list(np.logspace(0, 2, 100)),
         'fit_prior': [True, False]}

gsnb = GridSearchCV(mnb, param_grid=params, cv=5)

gsnb.fit(X_train, y_train)

gsnb.best_params_

gsnb.best_score_

y_pred_train_nb = gsnb.predict(X_train)
y_pred_test_nb = gsnb.predict(X_test)

train_score = accuracy_score(y_train, y_pred_train_nb)
test_score = accuracy_score(y_test, y_pred_test_nb)

print(f"Train accuracy: {round(train_score, 4)}")
print(f"Test accuracy: \t{round(test_score, 4)}")
